requirements.txt
langchain
openai
faiss-cpu
pypdf
sentence-transformers
streamlit
python-dotenv